+++ NEWS.md
@@ -9,6 +9,7 @@
 * As a consequence of #225 and #232, controllers no longer have a `unpopped()` method.
 * As another consequence of #225 and #232, `wait(mode = "all")` no longer guarantees that a task is available for `pop()`. It just consumes a condition variable signal. `pop()` should always be checked for `NULL` return values.
 * Remove the long-deprecated `promise()` method of the controller.
+* Deprecate the `processes` field of controllers. Asynchronous worker launch/termination operations are no longer supported (#237). Job arrays will be a better solution.
 
 # crew 1.2.1